+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-- [...]
+- **[FIX]** `Stepper` Add onChange callback for external components
 
 # v0.11.2 (27/09/2018)